Add method put  : /

+++ public/js/main.js
@@ -10,26 +10,25 @@ if (form) {
 
             for (let [name, value] of formData) {
 
-               if(typeof value === "object"){
-                   formData.append('nameImg', value.name)
-               }
-               data[name] = value;
+                if (typeof value === "object") {
+                    formData.append('nameImg', value.name);
+
+                }
+                data[name] = value;
             }
+        console.log(data);
 
 
         if (data.id) {
                 fetch('/notes', {
                     method: 'PUT',
-                    body: JSON.stringify(data),
-                    headers: {
-                        'Content-Type': 'application/json',
-                    }
+                    body: formData
                 }).finally(() => {
 
                     console.log("Go to space and Update!!!!!!");
 
                 });
-               //window.location.href = '/';
+                //window.location.href = '/';
 
             } else {
                 fetch('/notes', {
@@ -39,12 +38,12 @@ if (form) {
                     console.log(r);
 
                 }).finally(() => {
-                     window.location.href = '/';
+                    //  window.location.href = '/';
                     console.log("Go to space!!!!!!");
                 });
 
             }
-       // document.getElementById("img").setAttribute("src", `image/'${data.nameImg}'`);
+            // document.getElementById("img").setAttribute("src", `image/'${data.nameImg}'`);
 
         }
     );
@@ -57,41 +56,46 @@ function onBtnDeleteClick(event, id, imageId) {
     if (document.getElementById(id)) {
         document.getElementById(id).remove();
     } else {
-        window.location.href = '/';
+        // window.location.href = '/';
     }
 
-    console.log(imageId);
+    if (id !== null && id !== undefined && id.length > 0) {
+        fetch('/notes', {
+            method: 'DELETE',
+            body: JSON.stringify({id: id}),
+            headers: {
+                'Content-Type': 'application/json',
+            }
+        }).finally(() => {
+            console.log("Go to space and delete!!!!!!");
+        });
+    }
 
-    fetch('/notes', {
-        method: 'DELETE',
-        body: JSON.stringify({id: id}),
-        headers: {
-            'Content-Type': 'application/json',
-        }
-    }).finally(() => {
-        console.log("Go to space and delete!!!!!!");
-    });
+    if (imageId !== null && imageId !== undefined && imageId.length > 0) {
 
-    fetch('/notes/image', {
-        method: 'DELETE',
-        body: JSON.stringify({imageId: imageId}),
-        headers: {
-            'Content-Type': 'application/json',
-        }
-    }).finally(() => {
-        console.log("Go to space and delete!!!!!!");
-    });
+        fetch('/notes/image', {
+            method: 'DELETE',
+            body: JSON.stringify({imageId: imageId}),
+            headers: {
+                'Content-Type': 'application/json',
+            }
+        }).finally(() => {
+            console.log("Go to space and delete!!!!!!");
+        });
 
-
+    }
 }
 
 
-function onBtnChangeClick(e, id) {
+function onBtnChangeClick(e) {
 
-   // e.preventDefault();
-
-   let img = document.getElementById('fileChange');
-    return img.click()
+    e.preventDefault();
+    let changeImg = document.getElementById('fileChange');
+    changeImg.click();
 }
 
+function onFileChange(e, id, imageId) {
+
+    onBtnDeleteClick(e, id, imageId)
+}
